# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/faq-cache": {
      "get": {
        "operationId": "FaqCacheController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List cached FAQ entries",
        "tags": [
          "faq-cache"
        ]
      }
    },
    "/api/v1/faq-cache/{id}": {
      "get": {
        "operationId": "FaqCacheController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a cached FAQ entry",
        "tags": [
          "faq-cache"
        ]
      },
      "patch": {
        "operationId": "FaqCacheController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edit the answer of a cached FAQ entry",
        "tags": [
          "faq-cache"
        ]
      },
      "delete": {
        "operationId": "FaqCacheController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a cached FAQ entry",
        "tags": [
          "faq-cache"
        ]
      }
    },
...
  },
  "info": {
    "title": "Townibos API",
    "description": "Residential complex (perumahan) CRM — REST API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}